+++ js/bootstrap-3-shortcodes-for-wordpress.js
@@ -1,10 +1,8 @@
 (function($) {
-	$('[data-toggle="bootstrap-popover"]')
-		.on('click', function(e) {
-			e.preventDefault();
-			return true;
-		})
-		.popover();
+	$('[data-toggle="bootstrap-popover"]').on('click', function(e) {
+		e.preventDefault();
+		return true;
+	}).popover();
 
 	$('.bootstrap-tooltip').tooltip();
 })(jQuery);